# dist/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/api.localhost",
            "description": "ローカルサーバー"
        },
        {
            "url": "http:\/\/localhost:8000",
            "description": "モックサーバー"
        }
    ],
    "paths": {
...
        "\/v1\/auth\/reset": {
            "post": {
                "tags": [
                    "user_auth"
                ],
                "summary": "パスワードリセット",
                "description": "ユーザーのパスワードを再設定します。\nパスワードリセットメールに添付のURLのクエリパラメーター`email`と`token`を使用してください。",
                "operationId": "ResetPassword",
                "requestBody": {
                    "description": "パスワードリセット",
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/ResetPasswordParameter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/LoginUserTokenData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/BadRequestException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "title": "ValidationErrorResponse",
                                    "type": "object",
                                    "required": [
                                        "errors"
                                    ],
                                    "properties": {
                                        "errors": {
                                            "title": "ValidationErrors",
                                            "type": "object",
                                            "required": [
                                                "keys",
                                                "values"
                                            ],
                                            "properties": {
                                                "keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "エラーの対象リクエストキー名",
                                                        "type": "string",
                                                        "example": "email"
                                                    },
                                                    "example": [
                                                        "name",
                                                        "email"
                                                    ]
                                                },
                                                "values": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "エラーの内容",
                                                        "type": "string",
                                                        "example": "メールアドレスは必須です"
                                                    },
                                                    "example": [
                                                        "名前は255文字以内です",
                                                        "メールアドレスは必須です"
                                                    ]
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/TooManyRequestsException"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "\/v1\/admins\/auth\/login": {
            "post": {
                "tags": [
                    "admin_auth"
                ],
                "summary": "ユーザーログイン",
                "description": "ユーザーのログインです。\nアクセストークンの有効期限は60分です。",
                "operationId": "LoginAdmin",
                "requestBody": {
                    "description": "管理ユーザーログインパラメーター",
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/LoginAdminParams"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/LoginAdminTokenData"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/UnauthorizedException"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/TooManyRequestsException"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "\/v1\/admins\/auth\/reset": {
            "post": {
                "tags": [
                    "admin_auth"
                ],
                "summary": "パスワードリセット",
                "description": "ユーザーのパスワードを再設定します。\nパスワードリセットメールに添付のURLのクエリパラメーター`email`と`token`を使用してください。",
                "operationId": "AdminResetPassword",
                "requestBody": {
                    "description": "管理ユーザーパスワードリセット",
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/AdminResetPasswordParams"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/LoginAdminTokenData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/BadRequestException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "title": "ValidationErrorResponse",
                                    "type": "object",
                                    "required": [
                                        "errors"
                                    ],
                                    "properties": {
                                        "errors": {
                                            "title": "ValidationErrors",
                                            "type": "object",
                                            "required": [
                                                "keys",
                                                "values"
                                            ],
                                            "properties": {
                                                "keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "エラーの対象リクエストキー名",
                                                        "type": "string",
                                                        "example": "email"
                                                    },
                                                    "example": [
                                                        "name",
                                                        "email"
                                                    ]
                                                },
                                                "values": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "エラーの内容",
                                                        "type": "string",
                                                        "example": "メールアドレスは必須です"
                                                    },
                                                    "example": [
                                                        "名前は255文字以内です",
                                                        "メールアドレスは必須です"
                                                    ]
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/TooManyRequestsException"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}